+++ lending/RELEASES.md
@@ -539,3 +539,13 @@ Based on:
 - [typescript v0.6.2] lending
 ### Releases
 - [NPM v0.6.2] https://www.npmjs.com/package/ryan-lending/v/0.6.2 - lending
+
+## 2024-11-09 00:56:10
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.436.0 (2.455.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.6.3] lending
+### Releases
+- [NPM v0.6.3] https://www.npmjs.com/package/ryan-lending/v/0.6.3 - lending